Splitting the repositories into their own projects
