+++ src/NHilLo.Repository.SQLite/Class1.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace NHilLo.Repository.SQLite
+{
+    public class Class1
+    {
+
+    }
+}
