+++ src/NHilo.sln
@@ -16,6 +16,16 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		..\.github\workflows\dotnet.yml = ..\.github\workflows\dotnet.yml
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NHilLo.Repository.SQLServer", "NHilLo.Repository.SQLServer\NHilLo.Repository.SQLServer.csproj", "{874044AA-6953-499E-AB89-35A833FCF665}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NHilLo.Repository.Oracle", "NHilLo.Repository.Oracle\NHilLo.Repository.Oracle.csproj", "{565B7EF4-E579-4866-83D9-E9C5055A4D9B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NHilLo.Repository.MySql", "NHilLo.Repository.MySql\NHilLo.Repository.MySql.csproj", "{40DF5B22-66D7-4C76-ADF9-E633B4844942}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NHilLo.Repository.SQLite", "NHilLo.Repository.SQLite\NHilLo.Repository.SQLite.csproj", "{4B91F05A-F2E1-4360-9000-67A71A5BC789}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "repositories", "repositories", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -62,6 +72,54 @@ Global
 		{D240AC8B-5E41-44A6-B6F3-F71AC3ABF1F5}.Release|x64.Build.0 = Release|Any CPU
 		{D240AC8B-5E41-44A6-B6F3-F71AC3ABF1F5}.Release|x86.ActiveCfg = Release|Any CPU
 		{D240AC8B-5E41-44A6-B6F3-F71AC3ABF1F5}.Release|x86.Build.0 = Release|Any CPU
+		{874044AA-6953-499E-AB89-35A833FCF665}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{874044AA-6953-499E-AB89-35A833FCF665}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{874044AA-6953-499E-AB89-35A833FCF665}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{874044AA-6953-499E-AB89-35A833FCF665}.Debug|x64.Build.0 = Debug|Any CPU
+		{874044AA-6953-499E-AB89-35A833FCF665}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{874044AA-6953-499E-AB89-35A833FCF665}.Debug|x86.Build.0 = Debug|Any CPU
+		{874044AA-6953-499E-AB89-35A833FCF665}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{874044AA-6953-499E-AB89-35A833FCF665}.Release|Any CPU.Build.0 = Release|Any CPU
+		{874044AA-6953-499E-AB89-35A833FCF665}.Release|x64.ActiveCfg = Release|Any CPU
+		{874044AA-6953-499E-AB89-35A833FCF665}.Release|x64.Build.0 = Release|Any CPU
+		{874044AA-6953-499E-AB89-35A833FCF665}.Release|x86.ActiveCfg = Release|Any CPU
+		{874044AA-6953-499E-AB89-35A833FCF665}.Release|x86.Build.0 = Release|Any CPU
+		{565B7EF4-E579-4866-83D9-E9C5055A4D9B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{565B7EF4-E579-4866-83D9-E9C5055A4D9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{565B7EF4-E579-4866-83D9-E9C5055A4D9B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{565B7EF4-E579-4866-83D9-E9C5055A4D9B}.Debug|x64.Build.0 = Debug|Any CPU
+		{565B7EF4-E579-4866-83D9-E9C5055A4D9B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{565B7EF4-E579-4866-83D9-E9C5055A4D9B}.Debug|x86.Build.0 = Debug|Any CPU
+		{565B7EF4-E579-4866-83D9-E9C5055A4D9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{565B7EF4-E579-4866-83D9-E9C5055A4D9B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{565B7EF4-E579-4866-83D9-E9C5055A4D9B}.Release|x64.ActiveCfg = Release|Any CPU
+		{565B7EF4-E579-4866-83D9-E9C5055A4D9B}.Release|x64.Build.0 = Release|Any CPU
+		{565B7EF4-E579-4866-83D9-E9C5055A4D9B}.Release|x86.ActiveCfg = Release|Any CPU
+		{565B7EF4-E579-4866-83D9-E9C5055A4D9B}.Release|x86.Build.0 = Release|Any CPU
+		{40DF5B22-66D7-4C76-ADF9-E633B4844942}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{40DF5B22-66D7-4C76-ADF9-E633B4844942}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{40DF5B22-66D7-4C76-ADF9-E633B4844942}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{40DF5B22-66D7-4C76-ADF9-E633B4844942}.Debug|x64.Build.0 = Debug|Any CPU
+		{40DF5B22-66D7-4C76-ADF9-E633B4844942}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{40DF5B22-66D7-4C76-ADF9-E633B4844942}.Debug|x86.Build.0 = Debug|Any CPU
+		{40DF5B22-66D7-4C76-ADF9-E633B4844942}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{40DF5B22-66D7-4C76-ADF9-E633B4844942}.Release|Any CPU.Build.0 = Release|Any CPU
+		{40DF5B22-66D7-4C76-ADF9-E633B4844942}.Release|x64.ActiveCfg = Release|Any CPU
+		{40DF5B22-66D7-4C76-ADF9-E633B4844942}.Release|x64.Build.0 = Release|Any CPU
+		{40DF5B22-66D7-4C76-ADF9-E633B4844942}.Release|x86.ActiveCfg = Release|Any CPU
+		{40DF5B22-66D7-4C76-ADF9-E633B4844942}.Release|x86.Build.0 = Release|Any CPU
+		{4B91F05A-F2E1-4360-9000-67A71A5BC789}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4B91F05A-F2E1-4360-9000-67A71A5BC789}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4B91F05A-F2E1-4360-9000-67A71A5BC789}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{4B91F05A-F2E1-4360-9000-67A71A5BC789}.Debug|x64.Build.0 = Debug|Any CPU
+		{4B91F05A-F2E1-4360-9000-67A71A5BC789}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{4B91F05A-F2E1-4360-9000-67A71A5BC789}.Debug|x86.Build.0 = Debug|Any CPU
+		{4B91F05A-F2E1-4360-9000-67A71A5BC789}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4B91F05A-F2E1-4360-9000-67A71A5BC789}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4B91F05A-F2E1-4360-9000-67A71A5BC789}.Release|x64.ActiveCfg = Release|Any CPU
+		{4B91F05A-F2E1-4360-9000-67A71A5BC789}.Release|x64.Build.0 = Release|Any CPU
+		{4B91F05A-F2E1-4360-9000-67A71A5BC789}.Release|x86.ActiveCfg = Release|Any CPU
+		{4B91F05A-F2E1-4360-9000-67A71A5BC789}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -69,5 +127,12 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{2A7BB3EF-FCBC-44E9-824E-034CD4D4FC3E} = {CDE2D060-6E68-11EB-B827-83182767B3D8}
 		{D240AC8B-5E41-44A6-B6F3-F71AC3ABF1F5} = {CDE2D060-6E68-11EB-B827-83182767B3D8}
+		{874044AA-6953-499E-AB89-35A833FCF665} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{565B7EF4-E579-4866-83D9-E9C5055A4D9B} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{40DF5B22-66D7-4C76-ADF9-E633B4844942} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{4B91F05A-F2E1-4360-9000-67A71A5BC789} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F58ECF33-B291-4860-B339-BA894B58A03A}
 	EndGlobalSection
 EndGlobal
